+++ config/config.template.ini
@@ -88,7 +88,7 @@ httpRequest = 0
 eMail = 0
 BosMon = 0
 firEmergency = 0
-jsonUDP = 0
+jsonSocket = 0
 
 # for developing template-module
 template = 0
@@ -195,7 +195,9 @@ bosmon_password =
 firserver = localhost
 firport = 9001
 
-[jsonUDP]
+[jsonSocket]
+# Protocol for socket (TCP|UDP)
+protocol = UDP
 # Server as IP of DNS-Name (without http://)
 server = 192.168.0.1
 port = 8888
